src/poller/requirements.txt
requests==2.31.0
boto3-stubs[s3]==1.34.42
